moving from f64 to f32

+++ src/activation/mod.rs
@@ -1,6 +1,6 @@
 // activation functions
 
-use std::f64;
+use std::f32;
 use ndarray::{Array, Ix3};
 
 #[derive(Debug, PartialEq, Copy, Clone)]
@@ -13,25 +13,25 @@ pub enum ActivationFunctionType {
     Softmax,
 }
 
-pub fn sigmoid(x: &f64) -> f64 {
+pub fn sigmoid(x: &f32) -> f32 {
     let minus_x = -x;
     1.0 / (1.0 + minus_x.exp())
 }
 
-pub fn relu(x: &f64) -> f64 {
+pub fn relu(x: &f32) -> f32 {
     x.max(0.0)
 }
 
-pub fn leaky_relu(x: &f64, alpha: Option<f64>) -> f64 {
+pub fn leaky_relu(x: &f32, alpha: Option<f32>) -> f32 {
     let leak = alpha.unwrap_or_else(|| 0.1) * x;
     x.max(leak)
 }
 
-pub fn tanh(x: &f64) -> f64 {
+pub fn tanh(x: &f32) -> f32 {
     x.clone().tanh()
 }
 
-pub fn softmax(x: &Array<f64, Ix3>) -> Array<f64, Ix3> {
+pub fn softmax(x: &Array<f32, Ix3>) -> Array<f32, Ix3> {
     let exp_scores = x.clone().map(|x| x.exp());
     exp_scores.clone() / exp_scores.sum()
 }
