+++ src/layer/flatten.rs
@@ -14,7 +14,7 @@ impl Flatten {
         }
     }
 
-    fn forward_channels(&self, input: &Array<f64, Ix3>) -> Result<Array<f64, Ix3>, LayerError> {
+    fn forward_channels(&self, input: &Array<f32, Ix3>) -> Result<Array<f32, Ix3>, LayerError> {
         todo!()
         /*
         let input_shape = input.shape();
@@ -34,7 +34,7 @@ impl Flatten {
         */
     }
 
-    fn forward_with_no_channels(&self, input: &Array<f64, Ix3>) -> Result<Array<f64, Ix3>, LayerError> {
+    fn forward_with_no_channels(&self, input: &Array<f32, Ix3>) -> Result<Array<f32, Ix3>, LayerError> {
         let mut flatten_input = Array::from_iter(input.iter().cloned());
         let flatten_input_size = flatten_input.shape()[0];
         Ok(flatten_input.into_shape((1, flatten_input_size, 1)).unwrap().to_owned())
@@ -50,7 +50,7 @@ impl Layer for Flatten {
         ActivationFunctionType::None
     }
 
-    fn forward(&self, input: &Array<f64, Ix3>) -> Result<Array<f64, Ix3>, LayerError> {
+    fn forward(&self, input: &Array<f32, Ix3>) -> Result<Array<f32, Ix3>, LayerError> {
         if self.has_channels {
             return self.forward_channels(input)
         }
